// src/api/vcci-news/models/contact.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */

export interface Contact {
  id?: string;
  fullname?: string;
  email?: string;
  phone?: string;
  title?: string;
  content?: string;
  status?: string;
  created_at?: string;
  updated_at?: string;
  created_by?: string;
  updated_by?: string;
}